refactor: split flashcard-submit test file into focused modules
Split the large 766-line flashcard-submit.test.ts file into three focused test modules:
- flashcard-submit-db.test.ts: Database operations and core functionality (10 tests)
- flashcard-submit-success.test.ts: Success response validation (5 tests)
- flashcard-submit-error.test.ts: Comprehensive error handling (10 tests)

Improves test organization and maintainability while preserving all 25 original tests.
